Fix byte-order and type-width bugs the native boot exposed

+++ code/sha.h
@@ -31,6 +31,8 @@
 
 #pragma once
 
+#include <cstdint>
+
 
 #include <cstdio>
 #include <cstdlib>
@@ -72,7 +74,7 @@ class SHAEngine
 	private:
 
 		typedef union {
-			unsigned long Long[5];
+			uint32_t Long[5];
 			unsigned char Char[20];
 		} SHADigest;
